[sbserver] Refactor usageNanoCores be to used for all OSes
